# pessoas-api/swagger.yml
swagger: "2.0"

info:
  description: "Esse é o Swagger da API de Pessoa"
  version: "1.0.0"
  title: "Pessoa"

host: "127.0.0.1:3000"
basePath: "/"

tags:
- name: "pessoa"
  description: "Operações de pessoas"

schemes:
- "http"

paths:
  "/api/pessoas":

    get:
      tags:
      - "pessoa"

      summary: "Listar pessoas"
      description: "<h4>Requisição que irá buscar todas as pessoas da base de dados.</h4> "
      operationId: "listarPessoas"
      consumes:
      - "application/json"

      produces:
      - "application/json"

      parameters: []
      responses:
        200:
          description: Sucesso
          schema:
            $ref: "#/definitions/pessoaList"

          examples:
            response:
              - _id: 5bb04a60742e375c4a88edac
                name: Franck de Oliveira Aragão Júnior
                age: 22
                bithday: 27/11/1995
              - _id: 5bb04a60742e375c4a88edac
                name: Emerson Davi Alexandre dos Santos
                age: 22
                bithday: 07/05/1996
        500:
          description: Internal Server Error
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Internal Server Error:
              code: '500'
              message: Erro interno do servidor
    post:
      tags:
      - "pessoa"
      summary: "Cadastrar pessoa"
      description: "<h4>Requisição que irá realizar o cadastro de uma pessoa na base de dados.</h4>"
      operationId: "cadastrarPessoa"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Objeto do tipo Pessoa"
        required: true
        schema:
          $ref: "#/definitions/Pessoa"
      responses:
        200:
          description: Sucesso
          schema:
            $ref: "#/definitions/Pessoa"

          examples:
            response:
              - _id: 5bb04a60742e375c4a88edac
                name: Emerson Davi Alexandre dos Santos
                age: 22
                bithday: 07/05/1996
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Bad Request:
              code: '400'
              message: Requisição inválida
  "/api/pessoas/{pessoaId}":
    get:
      tags:
      - "pessoa"
      summary: "Buscar pessoa por ID"
      description: "<h4>Requisição que irá realizar a busca de uma pessoa por ID na base de dados.</h4>"
      operationId: "buscarPessoaPorId"
      produces:
      - "application/xml"
      - "application/json"

      parameters:
      - name: "pessoaId"
        in: "path"
        description: "ID da pessoa que deseja buscar na base de dados"
        required: true
        type: "string"

      responses:
        200:
          description: Sucesso
          schema:
            $ref: "#/definitions/Pessoa"
          examples:
            response:
            - _id: 5bb04a60742e375c4a88edac
              name: Emerson Davi Alexandre dos Santos
              age: 22
              bithday: 07/05/1996

        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ApiResponse"
          examples:
            Bad Request:
              code: '400'
              message: Requisição inválida

        404:
          description: Not Found
          schema:
            $ref: "#/definitions/ApiResponse"
          examples:
            Not Found:
              code: '404'
              message: Não encontrado
    delete:
      tags:
      - "pessoa"
      summary: "Deletar pessoa por ID"
      description: "<h4>Requisição que irá deletar uma pessoa da base de dados.</h4>"
      operationId: "deletarPessoa"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "pessoaId"
        in: "path"
        description: "ID da pessoa que deseja delatar da base de dados"
        required: true
        type: "string"
      responses:
        200:
          description: Sucesso

        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Bad Request:
              code: '400'
              message: Requisição inválida
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Not Found:
              code: '404'
              message: Não encontrado
    put:
      tags:
      - "pessoa"
      summary: "Atualizar pessoa"
      description: "<h4>Requisição que irá realizar a atualização das informações de uma pessoa na base de dados.</h4>"
      operationId: "atualizarPessoa"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "pessoaID"
        in: "path"
        description: "ID da pessoa que terá suas informações atualizadas"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Pessoa com as informações atualizadas para salvar na base de dados"
        required: true
        schema:
          $ref: "#/definitions/Pessoa"
      responses:
        200:
          description: Sucesso
          schema:
            $ref: "#/definitions/Pessoa"

          examples:
            response:
            - _id: 5bb04a60742e375c4a88edac
              name: Emerson Davi Alexandre da Silva
              age: 22
              bithday: 07/05/1996

        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Bad Request:
              code: '400'
              message: Requisição inválida
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/ApiResponse"

          examples:
            Not Found:
              code: '404'
              message: Não encontrado
definitions:
  pessoaList:
    type: array
    items:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        age:
          type: integer
        birthday:
          type: string
  Pessoa:
    type: object
    properties:
      name:
        type: string
      age:
        type: integer
      birthday:
        type: string

  ApiResponse:
    type: object
    properties:
      code:
        type: string
      message:
        type: string